Major Update - File Transfer Encryption
Files are now encrypted while being transferred.

File server is now started with the main program, instead of requiring running a separate script.

Current limit to files being sent is 4064 bytes. I am working on fixing this, the issue is that the nonce/tag gets cut off. I will be adding headers that specify the length of the sent data, which the receiver will use to ensure it gets the entire message (including nonce + tag).

The other annoying bug is that the file transfer notifications interfere with the cmd shell prompts. I will fix this if I have time, but nailing down file transfer is the priority.