Improved file transfer notifications
Now sends input request to main thread. Remaining issues are to trying to get the main thread to wait for the transfer to complete and then reprint the prompt, and getting the prompt to show up without needing to update the shell. Made an attempt at this and failed, not enough time to iron it out like I would like.